+++ src/index/stemmer-wasm/stemmer_wasm.d.ts
@@ -1,0 +1,4 @@
+/* tslint:disable */
+/* eslint-disable */
+
+export function stem(algorithm_name: string, word: string): string;